Refactor common lib
Issue: #2
