+++ source/CMakeLists.txt
@@ -1,3 +1,3 @@
-add_subdirectory(libcutter)
+add_subdirectory(textcutter-common)
 add_subdirectory(textcutter)
 add_subdirectory(textcutter-gui)
