+++ source/textcutter-common/common/text_cutter.h
@@ -1,0 +1,11 @@
+#ifndef COMMON_TEXT_CUTTER_H
+#define COMMON_TEXT_CUTTER_H
+
+#include <QtCore>
+
+namespace common
+{
+QVector<QString> CutText(const QString& text, const std::int32_t num_parts);
+}
+
+#endif // COMMON_TEXT_CUTTER_H
