+++ source/textcutter-gui/gui/main_window.cpp
@@ -1,5 +1,5 @@
 #include "gui/main_window.h"
-#include "cutter/cut_text.h"
+#include "common/text_cutter.h"
 
 namespace gui
 {
@@ -74,7 +74,7 @@ void MainWindow::CutText()
 {
     ClearPartTabs();
 
-    const auto parts{cutter::CutText(original_->toPlainText(), std::min(num_parts_->value(), kMaxNumberOfParts))};
+    const auto parts{common::CutText(original_->toPlainText(), std::min(num_parts_->value(), kMaxNumberOfParts))};
     const std::int32_t num_parts{std::min(parts.size(), kMaxNumberOfParts)};
 
     for (std::int32_t i{ 0 }; i < num_parts; ++i)
